+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Vanilla/default mods can now be prioritized over extended material mods
 - Mods can now be disabled such that assets from them are not used for patching your load order
 - "Use MO2 Loose File Order" option is now in the mod sort dialog
+- Removed optimize meshes post-patcher
 - Fixed duplicate mesh shapes not being patched for certain edge cases
 - Fixed null texture sets not being factored when reordering NIF blocks